# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de manipulação de produtos",
    "description": "Essa API tem como objetivo a manipulação dos produtos",
    "termsOfService": "http://localhost:3000/terms",
    "contact": {
      "email": "leonardo.thomaz@hotmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "API de test"
    }
  ],
  "paths": {
    "/products": {
      "post": {
        "summary": "Cadastro de produto",
        "description": "Essa rota será responsável por cadastrar um novo produto",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "Teclado Sorara Pro",
                    "description": "Teclado Sorara Pro Switch Optical Blue",
                    "price": 310
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Product Already Exists"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/products/findByName": {
      "get": {
        "description": "Busca de um produto pelo nome",
        "summary": "Busca de um produto pelo nome",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Nome do produto para busca",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Busca de um produto pelo id",
        "summary": "Busca de um produto pelo id",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto para busca",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Alterar o produto pelo id",
        "summary": "Alterar um produto pelo id",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto para alteração",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "Teclado Sorara Pro Single Color",
                    "description": "Teclado Sorara Pro Switch Optical Red",
                    "price": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "id": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}